Glacier Version 2
Glacier Version 2 takes the SWE from CemaNeige, and If SWE on the layer is smaller than 1mm and the temperature is above 0, then melt kicks in with now a fixed melt parameter.
Automatic calibration works with 4 param for GR4J and 2 param for CemaNeige but not with the additional melt parameter. For automatic calibration neede to adjust the Calibration:_Michel  function to add SnowPack in RunOptions$Outputs_Cal